Adds initial wip on synced lists
