+++ packages/core/__tests__/node.spec.ts
@@ -3,7 +3,6 @@ import {
   FormKitGroupValue,
   FormKitPlugin,
   FormKitNode,
-  bfs,
   resetCount,
 } from '../src/node'
 import { createConfig } from '../src/config'
@@ -34,49 +33,6 @@ describe('node', () => {
     expect(commitEvent).toHaveBeenCalledTimes(1)
     node.input(['a', 'b'], false)
     expect(commitEvent).toHaveBeenCalledTimes(2)
-  })
-
-  it('emits a singe commit event for type list', () => {
-    const commitEvent = vi.fn()
-    const lib = function libraryPlugin() {}
-    lib.library = (node: FormKitNode) => {
-      if (node.props.type === 'list') {
-        node.define({ type: 'list' })
-      } else if (node.props.type === 'group') {
-        node.define({ type: 'group' })
-      } else {
-        node.define({ type: 'input' })
-      }
-    }
-    const node = createNode({
-      props: { type: 'list' },
-      plugins: [lib],
-    })
-    node.on('commit', commitEvent)
-    const parentA = createNode({ props: { type: 'group' }, parent: node })
-    const parentB = createNode({ props: { type: 'group' }, parent: node })
-    const parentC = createNode({ props: { type: 'group' }, parent: node })
-    createNode({
-      name: 'a',
-      props: { type: 'text' },
-      parent: parentA,
-      value: undefined,
-    })
-    createNode({
-      name: 'b',
-      props: { type: 'text' },
-      parent: parentB,
-      value: undefined,
-    })
-    createNode({
-      name: 'c',
-      props: { type: 'text' },
-      parent: parentC,
-      value: undefined,
-    })
-    expect(commitEvent).toHaveBeenCalledTimes(9)
-    node.input([{}, {}, {}], false)
-    expect(commitEvent).toHaveBeenCalledTimes(13)
   })
 
   it('allows configuration to flow to children', () => {
@@ -136,65 +92,6 @@ describe('node', () => {
     expect(listenerB).toHaveBeenCalledTimes(1)
   })
 
-  it('does not emit config:{property} events when ancestors defines its own local value', () => {
-    const node = createNode({
-      // config: { locale: 'en' },
-      type: 'group',
-      children: [
-        createNode({
-          type: 'list',
-          name: 'list',
-          config: {
-            locale: 'fr',
-          },
-          children: [createNode()],
-        }),
-      ],
-    })
-    const listenerA = vi.fn()
-    const listenerB = vi.fn()
-    expect(node.at('list')?.props.locale).toBe('fr')
-    expect(node.at('list.0')!.props.locale).toBe('fr')
-    node.at('list')!.on('config:locale', listenerA)
-    node.at('list.0')!.on('config:locale', listenerB)
-    node.config.locale = 'zh'
-    expect(node.props.locale).toBe('zh')
-    expect(node.at('list')!.props.locale).toBe('fr')
-    expect(node.at('list.0')!.props.locale).toBe('fr')
-    expect(listenerA).toHaveBeenCalledTimes(0)
-    expect(listenerB).toHaveBeenCalledTimes(0)
-  })
-
-  it('can traverse into lists and groups', () => {
-    const group = createNode({
-      type: 'group',
-      children: [
-        createNode({ name: 'team' }),
-        createNode({
-          type: 'list',
-          name: 'users',
-          children: [
-            createNode({
-              type: 'group',
-              children: [
-                createNode({ name: 'email' }),
-                createNode({ name: 'password' }),
-              ],
-            }),
-            createNode({
-              type: 'group',
-              children: [
-                createNode({ name: 'email' }),
-                createNode({ name: 'password', value: 'foobar' }),
-              ],
-            }),
-          ],
-        }),
-      ],
-    })
-    expect(group.at('users.1.password')?.value).toBe('foobar')
-  })
-
   it('only traverses one layer deep when calling node.each', () => {
     const tree = createNode({
       type: 'group',
@@ -223,7 +120,7 @@ describe('node', () => {
     expect(callback).toHaveBeenCalledTimes(5)
   })
 
-  it('stops traversing nodes when node.walk callback returns false', () => {
+  it('stops traversing nodes when node.walk callback returns false and stopIfFalse is true', () => {
     const tree = createNode({
       type: 'group',
       children: [
@@ -235,6 +132,23 @@ describe('node', () => {
     const callback = vi.fn().mockReturnValue(false)
     tree.walk(callback, true)
     expect(callback).toHaveBeenCalledTimes(1)
+  })
+
+  it('stops traversing nodes when node.walk callback returns false', () => {
+    const tree = createNode({
+      type: 'group',
+      children: [
+        createNode({ name: 'a' }),
+        createNode({
+          type: 'group',
+          children: [createNode({ name: 'b' }), createNode({ name: 'c' })],
+        }),
+        createNode({ name: 'x' }),
+      ],
+    })
+    const callback = vi.fn().mockReturnValue(false)
+    tree.walk(callback)
+    expect(callback).toHaveBeenCalledTimes(5)
   })
 
   it('does not allow nodes of type input to be created with children', () => {
@@ -364,66 +278,6 @@ describe('node', () => {
     const child = createNode({ parent })
     expect(parent.children.length).toBe(2)
     expect(child.parent).toBe(parent)
-  })
-
-  it('can get a node’s index', () => {
-    const item = createNode()
-    createNode({
-      type: 'list',
-      children: [createNode(), createNode(), item, createNode()],
-    })
-    expect(item.index).toBe(2)
-  })
-
-  it('allows changing a node’s index by directly assigning it', () => {
-    const moveMe = createNode()
-    const parent = createNode({
-      type: 'list',
-      children: [createNode(), createNode(), moveMe, createNode()],
-    })
-    moveMe.index = 1
-    let children = [...parent.children]
-    expect(children[1]).toBe(moveMe)
-    moveMe.index = 3
-    children = [...parent.children]
-    expect(children[3]).toBe(moveMe)
-    moveMe.index = -1
-    children = [...parent.children]
-    expect(children[0]).toBe(moveMe)
-    moveMe.index = 99
-    children = [...parent.children]
-    expect(children[3]).toBe(moveMe)
-  })
-
-  it('can inject a new child directly into a parent at a given index', () => {
-    const list = createNode({
-      type: 'list',
-      children: [
-        createNode({ value: 'A' }),
-        createNode({ value: 'C' }),
-        createNode({ value: 'D' }),
-      ],
-    })
-    createNode({ value: 'B', parent: list, index: 1 })
-    expect(list.value).toStrictEqual(['A', 'B', 'C', 'D'])
-  })
-
-  it('can inject a new child directly into a parent at a given index and inherit the value', () => {
-    const A = createNode({ value: 'A' })
-    const C = createNode({ value: 'C' })
-    const list = createNode({
-      type: 'list',
-      children: [A, C],
-    })
-    const val = clone(list.value as string[])
-    val.splice(1, 0, 'B')
-    list.input(val, false)
-    expect(list.value).toStrictEqual(['A', 'B', 'C'])
-    const B = createNode({ value: undefined, parent: list, index: 1 })
-    expect(list.value).toStrictEqual(['A', 'B', 'C'])
-    expect(B.value).toBe('B')
-    B.input('Z', false)
-    expect(list.value).toStrictEqual(['A', 'Z', 'C'])
   })
 
   it('can inject a new value into a list without the list immediately inheriting that index’s sub values', () => {
@@ -1096,49 +950,6 @@ describe('value propagation in a node tree', () => {
     })
   })
 
-  it('can remove a child from the list’s values', async () => {
-    const food = createNode({
-      type: 'list',
-      children: [
-        createNode({ value: 'pizza' }),
-        createNode({ value: 'pasta' }),
-        createNode({ value: 'steak' }),
-        createNode({ value: 'fish' }),
-      ],
-    })
-    food.remove(food.at([2])!)
-    expect(food.children.length).toBe(3)
-    expect(food.isSettled).toBe(true)
-    expect(food.value).toStrictEqual(['pizza', 'pasta', 'fish'])
-  })
-
-  it('can remove a child from a list by destroying it', async () => {
-    const repeater = createNode({
-      type: 'list',
-      children: [
-        createNode({
-          type: 'group',
-          children: [createNode({ name: 'a', value: '123' })],
-        }),
-        createNode({
-          type: 'group',
-          children: [createNode({ name: 'a', value: 'abc' })],
-        }),
-        createNode({
-          type: 'group',
-          children: [createNode({ name: 'a', value: 'xyz' })],
-        }),
-      ],
-    })
-    const commitListener = vi.fn()
-    repeater.on('commit', commitListener)
-    repeater.at('1')?.destroy()
-    await repeater.settled
-    expect(repeater.children.length).toBe(2)
-    expect(repeater.value).toStrictEqual([{ a: '123' }, { a: 'xyz' }])
-    expect(commitListener).toHaveBeenCalledTimes(1)
-  })
-
   it('can remove a child from a group’s values', async () => {
     const address = createNode({
       type: 'group',
@@ -1316,36 +1127,6 @@ describe('value propagation in a node tree', () => {
     expect(treeA.at('b.d')!.value).toBe(456)
   })
 
-  it('can hydrate a list at depth', () => {
-    const tree = createNode({
-      type: 'group',
-      name: 'form',
-      value: {
-        a: 'foo',
-        people: ['first', 'second', 'third'],
-      },
-      children: [
-        createNode({ name: 'a' }),
-        createNode({
-          name: 'people',
-          type: 'list',
-          children: [
-            createNode(),
-            createNode({ value: 'fifth' }),
-            createNode(),
-          ],
-        }),
-      ],
-    })
-    expect(tree.value).toStrictEqual({
-      a: 'foo',
-      people: ['first', 'second', 'third'],
-    })
-    expect(tree.at('people.0')!.value).toBe('first')
-    expect(tree.at('people.1')!.value).toBe('second')
-    expect(tree.at('people.2')!.value).toBe('third')
-  })
-
   it('settles the group when a preserved input is removed', async () => {
     const group = createNode({ type: 'group' })
     const child = createNode({ parent: group, props: { preserve: true } })
@@ -1374,46 +1155,6 @@ describe('value propagation in a node tree', () => {
         { price: 5000, row: 'backstage' },
         { price: 200, seat: undefined },
       ],
-    })
-  })
-
-  describe('bfs', () => {
-    it('searches the parent node first', () => {
-      const parent = createNameTree()
-      expect(bfs(parent, 'tommy')).toBe(parent)
-    })
-
-    it('searches for a name in the children', () => {
-      const parent = createNameTree()
-      expect(bfs(parent, 'wendy')).toBe(parent.at('wendy'))
-    })
-
-    it('allows changing the searched property', () => {
-      const parent = createNameTree()
-      expect(bfs(parent, '555', 'value')).toBe(parent.at('jane'))
-    })
-
-    it('allows a callback to determine the search parameters', () => {
-      const parent = createNameTree()
-      expect(
-        bfs(
-          parent,
-          'radio',
-          (node) => node.name !== 'jane' && node.value === '555'
-        )
-      ).toBe(parent.at('stella.tommy'))
-    })
-
-    it('returns undefined when unable to find a match', () => {
-      const parent = createNameTree()
-      expect(bfs(parent, 'jim')).toBe(undefined)
-    })
-
-    it('searches the entire tree', () => {
-      const parent = createNameTree()
-      const searcher = vi.fn(() => false)
-      bfs(parent, 'jim', searcher)
-      expect(searcher.mock.calls.length).toBe(7)
     })
   })
