+++ packages/core/src/node.ts
@@ -9,6 +9,7 @@ import {
   init,
   cloneAny,
   clone,
+  isKeyedArray,
 } from '@formkit/utils'
 import {
   createEmitter,
@@ -394,6 +395,11 @@ export interface FormKitContext {
    */
   type: FormKitNodeType
   /**
+   * Only used on list nodes, this flag determines whether or not the list
+   * should sync its values with the underlying node children.
+   */
+  sync: boolean
+  /**
    * The actual value of the node.
    */
   value: unknown
@@ -551,18 +557,50 @@ export interface FormKitFrameworkContextState {
 }
 
 /**
- * Options that can be used to instantiate a new node via createNode().
+ * Options that can be used to instantiate a new node via `createNode()`.
  *
  * @public
  */
 export type FormKitOptions = Partial<
   Omit<FormKitContext, 'children' | 'plugins' | 'config' | 'hook'> & {
+    /**
+     * Config settings for the node, these are automatically exposed as props
+     * but are also checked in during hierarchical for prop checking.
+     */
     config: Partial<FormKitConfig>
+    /**
+     * Props directly set on this node, these are not inherited.
+     */
     props: Partial<FormKitProps>
+    /**
+     * The children of the node.
+     */
     children: FormKitNode[] | Set<FormKitNode>
+    /**
+     * The explicit index of this node when used in a list. If specified, this
+     * node will be created at this index atomically.
+     */
     index?: number
+    /**
+     * Should only be specified on list nodes — when true this indicates if the
+     * list node should automatically sync its child nodes with the value of
+     * the list node. In other words, if the list node’s value is an array of
+     * strings, and one string is popped off, the corresponding node should be
+     * removed the list and destroyed.
+     */
+    sync: boolean
+    /**
+     * Any plugins that should be registered on this node explicitly. These will
+     * automatically be inherited by any children.
+     */
     plugins: FormKitPlugin[]
+    /**
+     * For internal use only.
+     */
     alias: string
+    /**
+     * For internal use only.
+     */
     schemaAlias: string
   }
 >
@@ -1127,7 +1165,7 @@ export interface FormKitChildValue {
  *
  * @public
  */
-export type FormKitNode = {
+export type FormKitNode<V = unknown> = {
   /**
    * Boolean true indicating this object is a valid FormKitNode
    */
@@ -1136,7 +1174,7 @@ export type FormKitNode = {
    * The value of the input. This should never be directly modified. Any
    * desired mutations should be made through node.input()
    */
-  readonly value: unknown
+  readonly value: V
   /**
    * The internal FormKitContext object — this is not a public API and should
    * never be used outside of the core package itself. It is only here for
@@ -1722,7 +1760,8 @@ function hydrate(node: FormKitNode, context: FormKitContext): FormKitNode {
   const _value = context._value as KeyedValue
   context.children.forEach((child) => {
     if (typeof _value !== 'object') return
-    // if (has(context._value as FormKitGroupValue, child.name)) {
+    if (node.type === 'list' && node.sync)
+      return hydrateSyncedList(node, context)
     if (child.name in _value) {
       // In this case, the parent has a value to give to the child, so we
       // perform a down-tree synchronous input which will cascade values down
@@ -1755,6 +1794,27 @@ function hydrate(node: FormKitNode, context: FormKitContext): FormKitNode {
     }
   })
   return node
+}
+
+/**
+ * Hydrate a list node and its children
+ * @param node - A {@link FormKitNode | FormKitNode}
+ */
+function hydrateSyncedList(node: FormKitNode, context: FormKitContext) {
+  const _value = node._value
+  if (isKeyedArray(_value)) {
+  } else if (Array.isArray(_value)) {
+    let i = 0
+    for (; i < _value.length; i++) {
+      const child = node.children[i]
+      if (!eq(child._value, _value[i])) {
+        child.input(_value[i], false)
+      }
+    }
+    if (i < node.children.length) {
+      context.children.splice(i).forEach((child) => child.remove())
+    }
+  }
 }
 
 /**
@@ -2472,7 +2532,7 @@ function createConfig(
           node.emit(`config:${prop}`, value, false)
           configChange(node, prop, value)
           // Walk the node tree and notify of config/prop changes where relevant
-          node.walk((n) => configChange(n, prop, value), true)
+          node.walk((n) => configChange(n, prop, value))
         }
         return didSet
       }
@@ -2716,6 +2776,7 @@ function createContext(options: FormKitOptions): FormKitContext {
     props: createProps(value),
     settled: Promise.resolve(value),
     store: createStore(true),
+    sync: options.sync || false,
     traps: createTraps(),
     type: options.type || 'input',
     value,
@@ -2732,7 +2793,10 @@ function createContext(options: FormKitOptions): FormKitContext {
  *
  * @internal
  */
-function nodeInit(node: FormKitNode, options: FormKitOptions): FormKitNode {
+function nodeInit<V>(
+  node: FormKitNode,
+  options: FormKitOptions
+): FormKitNode<V> {
   // Set the internal node on the props, config, ledger and store
   node.ledger.init((node.store._n = node.props._n = node.config._n = node))
   // Apply given in options to the node.
@@ -2771,7 +2835,7 @@ function nodeInit(node: FormKitNode, options: FormKitOptions): FormKitNode {
   // Our node is finally ready, emit it to the world
   node.emit('created', node)
   node.isCreated = true
-  return node
+  return node as FormKitNode<V>
 }
 
 /**
@@ -2797,7 +2861,9 @@ function nodeInit(node: FormKitNode, options: FormKitOptions): FormKitNode {
  *
  * @public
  */
-export function createNode(options?: FormKitOptions): FormKitNode {
+export function createNode<V = unknown>(
+  options?: FormKitOptions
+): FormKitNode<V> {
   const ops = options || {}
   const context = createContext(ops) as FormKitContext
   // Note: The typing for the proxy object cannot be fully modeled, thus we are
